requirements-minimal.txt
# Minimal Requirements - Most Compatible Versions
# Use this if the main requirements.txt has conflicts

# Core ML (older but stable versions)
torch==1.13.1
torchvision==0.14.1
tensorflow==2.11.0
mediapipe==0.9.3.0

# Computer Vision
opencv-python==4.8.1.78
Pillow==9.5.0

# Scientific Computing
numpy==1.21.6
matplotlib==3.6.3

# Utilities
tqdm==4.64.1

# Dependencies
absl-py==1.4.0
protobuf==3.20.3
